backend/requirements.txt
fastapi
uvicorn
pandas
openai
anthropic
python-multipart
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
cryptography
PyMuPDF
openpyxl
